% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gui.R
\name{ac_gui}
\alias{ac_gui}
\title{REMS to Aquachem GUI
Shiny app to launch an interactive interface for getting and formating data.}
\usage{
ac_gui()
}
\description{
REMS to Aquachem GUI
Shiny app to launch an interactive interface for getting and formating data.
}
\examples{
\dontrun{ac_gui()}

}
