% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{check_present}
\alias{check_present}
\title{Check for data in `rems`}
\usage{
check_present(d, ems_ids, type = "all")
}
\arguments{
\item{d}{Data frame. Filtered output from [get_rems()].}

\item{ems_ids}{Character vector. EMS IDs}

\item{type}{Character. Either "all" (all data) or "date" (date-filtered data)}
}
\value{
Character vector of EMS IDs that __were__ in the data (omitting ones
  that were not found).
}
\description{
Returns a message regarding whether the data requested was actually found.
}
\keyword{internal}
