% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_rems}
\alias{get_rems}
\title{Fetch data from rems}
\usage{
get_rems(ems_ids, date_range, interactive)
}
\arguments{
\item{ems_ids}{Character vector. EMS IDs to filter by.}

\item{date_range}{Character vector. Start and end of the date range
(YYYY-MM-DD)}

\item{interactive}{Logical. Whether or not to allow interactive queries by
the `rems` package.}
}
\value{
data frame
}
\description{
Use [rems::rems-package()] to download data and then filter it.
}
\keyword{internal}
