% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gui.R
\name{gw_app}
\alias{gw_app}
\title{Launch Shiny app}
\usage{
gw_app()
}
\description{
Interactive interface for filtering and formatting data, summarizing water
quality and creating piper and stiff plots.
}
\examples{
\dontrun{gw_app()}

}
\concept{Shiny App User-Interface}
