% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meq}
\alias{meq}
\title{Calculate MEQ values}
\usage{
meq(d, drop_na = FALSE)
}
\arguments{
\item{d}{Data frame. Long data containing parameters and results}

\item{drop_na}{Logical. Whether to omit missing parameters}
}
\description{
Calculates MEQ values for 'long' data in mg/L. Expects columns:
"aqua_code" and "RESULT", where "aqua_code" is the parameter type
(e.g., "Zn_diss") and RESULT is the numeric concentration in mg/L.
}
\details{
For conversion details see the included data frame, \code{meq_conversion}.

Also see \code{?meq_conversion} for a description of the data.

MEQs are calculated by dividing the parameter concentration in mg/L by the
conversion factor.

\strong{Note: This is an internal function, exported for clarity in calculations}
}
\examples{

d <- data.frame(aqua_code = c("Cl", "HCO3"), RESULT = c(5.7, 38.3))
d
meq(d, drop_na = TRUE)
meq(d)


}
