% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_quality.R
\name{water_quality}
\alias{water_quality}
\title{Assess problems with water quality}
\usage{
water_quality(d)
}
\arguments{
\item{d}{EMS data formated for AquaChem as returned by \code{rems_to_aquachem()}}
}
\value{
Tibble of sites, samples, parameters, water quality limits and problems.
}
\description{
Assess problems with water quality
}
