% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{charge_balance}
\alias{charge_balance}
\title{Calculate charge balance}
\usage{
charge_balance(d)
}
\arguments{
\item{d}{AquaChem formatted dataset}

\item{return}{Return "all" columns or only "relevant" columns?}
}
\value{
Data frame
}
\description{
Calculate charge balance
}
