% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{charge_balance}
\alias{charge_balance}
\title{Calculate charge balance}
\usage{
charge_balance(d)
}
\arguments{
\item{d}{Data set formatted for AquaChem (output of \code{rems_to_aquachem()})}
}
\value{
Data frame
}
\description{
Calculates charge balances based on ALS formula. \strong{Note:} Original EMS
charge balances, anion sums and cation sums have been \strong{omitted}.
}
\details{
Potential changes in workflows over the years have made it difficult to
ascertain exactly how charge balances were calculated in older samples. This
resulted in discrepancies between EMS and locally calculated charge balances.
Therefore for consistency, we calculate charge balances for all samples using
the ALS formula below.

One difference between this calculation and that of ALS, is that we use more
significant digits when calculating MEQ.

anion sum = Cl_meq + SO4_meq + F_meq + NO3_meq + NO2_meq + Means_Alk_meq

cation sum = Ca_meq + Mg_meq + Na_meq + K_meq + Al_diss_meq +
Cu_diss_meq + Fe_diss_meq + Mn_diss_meq + Zn_diss_meq + NH4_meq +
(10 ^ (-pH_lab)) * 1000

Charge balance = 100 x (Cation Sum - Anion sum) / (Cation Sum + Anion Sum)

Missing values are ignored (ie. generally treated as 0). However, if all
values for cations or anions are missing the charge balance is \code{NA}.
}
