% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dominant_water_types}
\alias{dominant_water_types}
\title{Find the dominant water types}
\usage{
dominant_water_types(d, n = 5, p = 0.75)
}
\arguments{
\item{n}{Numeric. Number of samples above which to remove outliers.}

\item{p}{Numeric. Proportion of data required to assign a dominant water
type.}

\item{ems}{Data frame. EMS data for a single site.}
}
\value{
ems data frame with added \code{dominant} column (TRUE or FALSE) if the
water type is in the dominant set.
}
\description{
Find the dominant waters type in a group of samples, if there are more than
\code{n} samples. Match HCO3 and HCO3* types. Considered dominant types if in >\code{p}
of samples.
}
