% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{piper_plot}
\alias{piper_plot}
\title{Create Piper plot}
\usage{
piper_plot(
  d,
  ems_id = NULL,
  point_size = 0.1,
  colour = "ems_id",
  legend = TRUE,
  valid = TRUE,
  plot_data = FALSE,
  legend_position = "ul"
)
}
\arguments{
\item{d}{Data frame. AquaChem formatted dataset}

\item{ems_id}{Character. Ids to plot if dataset includes more than one}

\item{point_size}{Numeric. Point size}

\item{colour}{Character. Which column colour points by.}

\item{legend}{Logical. Whether to show the legend}

\item{valid}{Logical. Keep only valid data (charge balances <=10)}

\item{plot_data}{Logical. Whether to return plot data rather than a plot}

\item{legend_position}{Character. Location of legend. Must be one of "ul",
"ur", "ll", "lr", "cl", "cr", "uc", "lc", or "cc" for "upper left", "upper
right", etc.}
}
\description{
Create Piper plot
}
