% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{piper_plot}
\alias{piper_plot}
\title{Create Piper plot}
\usage{
piper_plot(d, ems_id = NULL, point_size = 0.1, colour = TRUE, legend = TRUE)
}
\arguments{
\item{d}{AquaChem formatted dataset}

\item{ems_id}{Ids to plot if dataset includes more than one}
}
\description{
Create Piper plot
}
