% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{piper_plot}
\alias{piper_plot}
\title{#' Calculate charge balance
#'
#' @param d AquaChem formatted dataset
#' @param return Return "all" columns or only "relevant" columns?
#'
#' @return Data frame
#' @export
#'
charge_balance <- function(d) {
  d %>%
    dplyr::mutate(cations = .data$Ca_meq + .data$Mg_meq + .data$Na_meq + .data$K_meq,
                  anions = .data$Cl_meq + .data$HCO3_meq + .data$SO4_meq,
                  charge_balance = ((.data$cations - abs(.data$anions)) /
                                      (.data$cations + abs(.data$anions))) * 100)
}
Create Piper plot}
\usage{
piper_plot(d, ems_id = NULL, point_size = 0.1, legend = TRUE)
}
\arguments{
\item{d}{AquaChem formatted dataset}

\item{ems_id}{Ids to plot if dataset includes more than one}
}
\description{
#' Calculate charge balance
#'
#' @param d AquaChem formatted dataset
#' @param return Return "all" columns or only "relevant" columns?
#'
#' @return Data frame
#' @export
#'
charge_balance <- function(d) {
  d %>%
    dplyr::mutate(cations = .data$Ca_meq + .data$Mg_meq + .data$Na_meq + .data$K_meq,
                  anions = .data$Cl_meq + .data$HCO3_meq + .data$SO4_meq,
                  charge_balance = ((.data$cations - abs(.data$anions)) /
                                      (.data$cations + abs(.data$anions))) * 100)
}
Create Piper plot
}
