% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{piper_plot}
\alias{piper_plot}
\title{Create Piper plot}
\usage{
piper_plot(
  d,
  ems_id = NULL,
  group = "ems_id",
  legend = TRUE,
  legend_position = "topleft",
  legend_title = group,
  valid = TRUE,
  plot_data = FALSE,
  point_colour = "viridis",
  point_size = 0.1,
  point_filled = TRUE,
  point_shape = "circle"
)
}
\arguments{
\item{d}{Data frame. AquaChem formatted dataset}

\item{ems_id}{Character. Ids to plot if dataset includes more than one}

\item{group}{Character. Column by which to group data for colour, shape,
filled and size.}

\item{legend}{Logical. Whether to show the legend}

\item{legend_position}{Character or Numeric.. Location of legend. Must be one
of "topleft", "topright", etc. (see ?legend for more options), OR a vector
of two numeric values x, and y to specify an exact position.}

\item{legend_title}{Character. Title of legend. Defaults to \code{group}.}

\item{valid}{Logical. Keep only valid data (charge balances <=10)}

\item{plot_data}{Logical. Whether to return plot data rather than a plot}

\item{point_colour}{Character. Colour or colours by which to colour points.
Either a single value (applied to all), or a vector of values the same
length as the number of \code{groups}. Can also be "viridis", which will use the
viridis colour scale.}

\item{point_size}{Numeric. Point size. Either a single value (applied to
all), or a vector of values the same length as the number of \code{groups}.}

\item{point_filled}{Logical. Whether to fill point shapes or not. Either a
single value (applied to all), or a vector of values the same length as the
number of \code{groups}.}

\item{point_shape}{Character. Shape of points to use. Valid options are
"circle", "square" or "triangle". Either a single value (applied to all),
or a vector of values the same length as the number of \code{groups}.}
}
\description{
Create Piper plot
}
