% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stiff_plot}
\alias{stiff_plot}
\title{Create Stiff plot}
\usage{
stiff_plot(d, ems_id = NULL, colour = TRUE, legend = TRUE, valid = TRUE)
}
\arguments{
\item{d}{AquaChem formatted dataset}

\item{ems_id}{Ids to plot if dataset includes more than one}

\item{colour}{Whether to add colour by ems_id}

\item{legend}{Whether to show the legend}

\item{valid}{Logical. Keep only valid data (charge balances <=10)}
}
\description{
Create Stiff plot
}
