% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stiff_plot}
\alias{stiff_plot}
\title{Create Stiff plot}
\usage{
stiff_plot(d, ems_id = NULL)
}
\arguments{
\item{d}{AquaChem formatted dataset}

\item{ems_id}{Ids to plot if dataset includes more than one}
}
\description{
Create Stiff plot
}
