% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{water_type}
\alias{water_type}
\title{Calculate water type}
\usage{
water_type(d)
}
\arguments{
\item{d}{Data frame. Must contain columns \code{Sample_Date}, \code{SampleID},
\code{StationID}, \code{Cl_meq}, \code{SO4_meq}, \code{HCO3_meq}, \code{Ca_meq}, \code{Mg_meq}, \code{Na_meq},
\code{K_meq}, and \code{charge_balance}.}
}
\value{
Data frame with added column \code{water_type}.
}
\description{
Water type based on anions Cl, SO4, HCO3 and cations Ca, Mg, Na and K. Elements
are ranked by proportion MEQ, all greater than 10\% are listed in descending
order of presence, cations first. Water type is only calculated for samples
with valid charge balances. Missing ions are ignored (i.e. treated as 0).
}
\examples{

d <- data.frame(Sample_Date = "2022-01-01", SampleID = "999990-01", StationID = 000,
                Cl_meq = 0.0226, SO4_meq = 0.0208, HCO3_meq = 1.54,
                Ca_meq = 0.187, Mg_meq = 0.490, Na_meq = 0.465, K_meq = 0.0665,
                charge_balance = 0.5)

d <- water_type(d)
d

}
